Backup-tool: Initial import
Restore backups created from VolumeCare in CloudStack

backup-tool.py implements 3 functions:
    - list available backups (for a given vm)
    - revert an existing VM (all volumes)
    - restore volume from backup and attach it to another VM

This version of backup-tool requires VolumeCare to have `id_tag=uuid`
in the configuration. It exits with an error if the backup was made
with the old format - i.e. without id_tag=uuid.

In the new format VolumeCare stores the volume UUID on each snapshot.
This is important if the CS volume path has changed between the backup and
the restore. This can happen if a DR sequence is executed for the VM, of
for other reason storpool volume gID has been changed.

Other imporovements:
    - list volumes uuid for each backup
    - supports future versions of VolumeCare where uuid will not be
      prefixed with ~
    - exits with an error if a backup with gID is found
      (made without id_tag=uuid)
    - README added