backup-tool: create the new volume with the VM owner
On attach command, the new volume is created in the same zone, domain
and user as the VM. This is requied to allow the volume to be attached
to the VM.

This allows the volumes to be attached to VMs owned by different owners.

The API key needs to have admin privileges.